Move state diff to state module
